Refactoring

+++ app/src/main/java/hu/gyuriczaadam/sprintformteszt/domain/model/TransactionItem.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 data class TransactionItem(
     val summary: String,
     val paid: String,
-    val sum: Int,
+    val sum: Long,
     val currency: String,
     val id:Int,
     val type:String,
